requirements.txt
absl-py==1.2.0
anyio==3.6.1
astunparse==1.6.3
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.1
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.81.0
flatbuffers==20181003210633
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3
grpcio==1.48.1
gTTS==2.2.4
h11==0.13.0
h5py==3.7.0
httplib2==0.20.4
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
Jinja2==3.1.2
keras==2.10.0rc1
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mtcnn==0.1.1
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
playsound==1.3.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.1
pyparsing==3.0.9
pyserial==3.5
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
starlette==0.19.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0rc3
tensorflow-cpu-aws==2.10.0rc3
tensorflow-estimator==2.10.0rc0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
typing-extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
